refactor structura of pages
